+++ qa/integration-host.cjs
@@ -2,4 +2,6 @@
 // Windows validation; mock those OS integrations, retaining production IPC.
 const {app}=require('electron');
 app.requestSingleInstanceLock=()=>true;
+app.getVersion=()=>require('../package.json').version;
+global.__logicflowWorker=require('../src/worker.cjs').Worker;
 require('../src/main.cjs');